Update DataReader.java
Added a sorting step after reading all the data from the CSV file. This comparator first sorts the data by location (country or continent) and then by date. This ensures that the data is always in the correct order, regardless of how it was originally arranged in the CSV file. This sorting logic addresses the assumption that the CSV file is already sorted.